+++ Rendering/Model.h
@@ -19,6 +19,6 @@ private:
 
 	std::vector<glm::vec3> _vertices, _normals;
 	std::vector<glm::vec2> _uvs;
-	std::vector<unsigned short> _indices;
+	std::vector<unsigned int> _indices;
 };
